:bug: Ensure that system check dependencies are loaded
System checks are loaded in the registry as a set datastructure,
which has no inherent ordering. It's possible that the admin
patching check runs before the URL conf was loaded, which leads
to false positives.

Instead, invoke django's core URL check to ensure that the
urlconf is loaded which should trigger the necessary side
effects.