:sparkles: [#9] Implement hijack totp-device management
Thanks to @Bartvaderkin for the original implementation

Added signals so that hijacking admin users works without
entering their second factor by (OTP) logging them in with
a temporary device.

Special care is taken to ensure hijacking is only allowed
for users that are verified, even if you forget to specify
a custom permission check. This functionality explicitly
tests that it cannot be abused to circumvent multi-factor
authentication.